back_end/python_service/requirements.txt
Django==4.2.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==21.2.0